target/stdbu: addition of ST DBU target for ST STi SOC
Addition of a ST DBU dedicated target here in order to load
the BTFMW faster via JTAG.  Benefit of having a target is
also to be able to use image loading functions of openocd
and thus being able to load directly from the ELF file of
the BTFMW.  Only few commands are supported, read and write
while write is hightly optimized (too much ?) is order to
load the BTFMW rapidly.

I might have to rework this since while writing this I was
using a "bad ?" USB Blaster which apparently ran very slowly,
leading to needs to optimise A LOT in order to achieve
reasonable loading time.  I now use a FTDI 2232H based JTAG
probe which behave way faster, so such optimization might not
be necessary and it might be better to behave more safely when
writing data.

Signed-off-by: Alain Volmat <avolmat@me.com>