tcl: addition of stih418 target and 4kopen_b2264 board
Addition of the stih418 target which can be found on the 4kopen board
(aka B2264).  This work is based on intensive reading of the .py files
found within the targetpack package available on 4kopen.com and used
to initialize the B2264 / STiH418 SoC over JTAG.  With that it is
possible to start the SoC and have the BTFWM binary loaded over JTAG
and running in order to later on connect to the A9 cores in order to
load some programs.

Tested with the following openocd command:
./src/openocd -s tcl -f tcl/board/4kopen_b2264.cfg -c init -c "debug_cold path_to_btfwm"

Both BTFWM and BTFWM params are needed for this and can be found within
the targetpack-20160812-3 found on 4kopen.com.  Following files should
be renamed / stored locally and path given to the debug_cold proc of the
target:
targetpack-20160812-3/boards_ng/b2264stxh418/btfmw-h418.out ->
btfmw-h418.elf
targetpack-20160812-3/boards_ng/b2264stxh418/btfmw_params-b2264stxh418rev1.bin

Signed-off-by: Alain Volmat <avolmat@me.com>